--- app/javascript/src/components/Dashboard/Notes/Description.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-import { Typography } from "neetoui";
-
-const Description = ({ description }) => (
-  <div className="mb-4 flex text-left">
-    <Typography style="body3">{description}</Typography>
-  </div>
-);
-export default Description;
