+++ app/javascript/src/components/Dashboard/Notes/List.jsx
@@ -2,11 +2,11 @@ import React from "react";
 
 import Note from "./Note";
 
-const Box = ({ fetchNotes, notes }) => (
+const List = ({ fetchNotes, notes }) => (
   <div className="flex w-full flex-col">
     {notes.map(note => (
       <Note fetchNotes={fetchNotes} key={note.id} note={note} />
     ))}
   </div>
 );
-export default Box;
+export default List;
