+++ app/javascript/src/components/Dashboard/Notes/Note.jsx
@@ -1,15 +1,71 @@
-import React from "react";
+import React, { useState } from "react";
+
+import { MenuVertical, Clock } from "neetoicons";
+import { Tooltip, Typography, Dropdown, Avatar, Button } from "neetoui";
 
 import Container from "./Container";
-import Description from "./Description";
-import Footer from "./Footer";
-import Header from "./Header";
+import DeleteAlert from "./DeleteAlert";
+import { calculateCreatedAgo, formatDateToWeekDay } from "./utils";
 
-const Note = ({ fetchNotes, note }) => (
-  <Container key={note.id}>
-    <Header fetchNotes={fetchNotes} note={note} />
-    <Description description={note.description} />
-    <Footer note={note} />
-  </Container>
-);
+const Note = ({ fetchNotes, note }) => {
+  const [showDeleteAlert, setShowDeleteAlert] = useState(false);
+  const [selectedNote, setSelectedNote] = useState({});
+  const [selectedNoteIds, setSelectedNoteIds] = useState([]);
+
+  const handleDelete = () => {
+    setShowDeleteAlert(true);
+    setSelectedNote(note);
+    setSelectedNoteIds([note.id]);
+  };
+
+  return (
+    <Container key={note.id}>
+      <div className="flex justify-between text-xl font-bold">
+        <Typography style="h3">{note.title}</Typography>
+        <Dropdown buttonStyle="text" icon={MenuVertical}>
+          <li>Edit</li>
+          <li onClick={handleDelete}>Delete</li>
+        </Dropdown>
+      </div>
+      {showDeleteAlert && (
+        <DeleteAlert
+          refetch={fetchNotes}
+          selectedNote={selectedNote}
+          selectedNoteIds={selectedNoteIds}
+          setSelectedDeleteNote={setSelectedNote}
+          onClose={() => setShowDeleteAlert(false)}
+        />
+      )}
+      <div className="mb-4 flex text-left">
+        <Typography style="body3">{note.description}</Typography>
+      </div>
+      <div className="flex items-center justify-between">
+        <Button
+          className="mx-2 mt-3"
+          label="Getting Started"
+          style="secondary"
+        />
+        <span className="flex justify-between">
+          <Clock className="mx-1 mt-3" />
+          <Tooltip
+            content={formatDateToWeekDay(note.created_at)}
+            followCursor="horizontal"
+            position="bottom"
+          >
+            <Typography className="mx-1 mt-4" style="body3">
+              Created {calculateCreatedAgo(note.created_at)}
+            </Typography>
+          </Tooltip>
+          <Avatar
+            className="mx-1 mt-3"
+            size="small"
+            user={{
+              imageUrl: "https://i.pravatar.cc/300",
+            }}
+          />
+        </span>
+      </div>
+    </Container>
+  );
+};
 export default Note;
